MDL-46116 mod_imscp: Better navigation of CP items with no resource attached.
Resource-less items are used as containers, and do not have any content of
their own to be displayed.
On CP load, display the first item with a resource, rather than just the first
child, which could be a container.
Selecting a container will update the content area with the item's title and
links to its children.